app/requirements.txt
flask
prometheus_flask_exporter
numpy
torch
Pillow
torchvision